refactor: cloudinary folder
